requirements.txt

beautifulsoup4
docopt
Jinja2
Keras
matplotlib
msgpack-python
numpy
pandas
Scrapy
tensorflow
urllib3